Binary: add Reader
If Value is nil and Reader is not, then Reader is used.
The bindReader is a little bit complicated, as only three types of piece
sequences are allowed:
    a) OCI_ONE_PIECE, one chunk
    b) OCI_FIRST_PIECE, OCI_LAST_PIECE (two, non-empty chunks)
    c) OCI_FIRST_PIECE, OCI_NEXT_PIECE*, OCI_LAST_PIECE
None of the chunks can be empty, so we have to pre-read the next chunk,
before sending the actual, to know whether this is the last or not.