Another take on Issue #36
1. factored the logic into a separate function at the end of rsetCfg.go,
named numericColumnType, so it will be easier to test it, once we
figured out all the corner cases and what to do in each.

2. modified TestUnderflow to create a table with all the possible number
column types (AFAIK), and create a view on it, if possible.

Still can't reproduce, but obviously my DB&table is not the same as the
one with the problem, so I expect this to still fail in some
environments.